+++ general/strFunc.h
@@ -37,7 +37,7 @@ bool isNo(char* answer);
 
 size_t utf8StrLen(char* str);
 void utf8Shift(int amountSymShift, char** pos);
-int getUtf8CharLength(char c);
+size_t getUtf8CharLength(char c);
 void skipWhitespace(char** bufferPos);
 bool isApprovedVariableSym(const char* str);
 bool isSpecialVarSymUtf8(const char* str);
